Shuffle code around between providers
