+++ app/webpacker/components/CompetitionForm/Tables/SeriesComps.js
@@ -9,8 +9,7 @@ import Loading from '../../Requests/Loading';
 import TableWrapper from './TableWrapper';
 import { useStore } from '../../../lib/providers/StoreProvider';
 import useLoadedData from '../../../lib/hooks/useLoadedData';
-import { useFormUpdateAction } from '../../wca/FormBuilder/EditForm';
-import { useFormObject } from '../../wca/FormBuilder/provider/FormObjectProvider';
+import { useFormObject, useFormUpdateAction } from '../../wca/FormBuilder/provider/FormObjectProvider';
 
 function MissingInfo({ missingDate, missingLocation }) {
   return (
